Now atleast it doesn't get error if storage dir not exist

+++ app/Helper/Viewer/Parser.php
@@ -4,12 +4,24 @@ namespace App\Helper\Viewer;
 use App\Helper\Router\Route;
 
 class Parser {
-	protected $fileContent;
+	static $fileContent;
+	static $compiledLocation = __DIR__ . "/../../../storage/views/";
+	static $srcLocation = __DIR__ . "/../../../resources/views/";
 	function __construct() {
 		echo "Hi";
 	}
 
-	static function Parse(string $fileLocation, string $viewName) {
+	static function fExist() {
+		if (file_exists(self::$compiledLocation)) {
+			return true;
+		} else {
+			mkdir(self::$compiledLocation, 0777, true);
+			return true;
+		}
+		return false; 
+	}
+
+	static function Parse(string $viewName, string $fileLocation) {
 		$fileName = $fileLocation . $viewName;
 		$fileContent = '';
 		if (str_contains($viewName, "temp.php")) {
@@ -47,11 +59,18 @@ class Parser {
 			$viewName = str_replace('.temp.php', '.php', $viewName);
 			echo $fileLocation . "COMPILED_" . $viewName;
 			$fileName = $fileLocation . "COMPILED_" . $viewName;
+			if (!self::fExist()) {
+				die();
+			}
 			$srcFileTime = filemtime($srcFileName);
+			if (file_exists(self::$compiledLocation . $viewName)) {
 			$fileTime = filemtime($fileName);
+			} else {
+				touch($fileName);
+				$fileTime = true;
+			}
 			$diffTime = $fileTime - $srcFileTime;
 			if (!$diffTime < 1000) {
-				touch($fileName);
 				file_put_contents($fileName, $fileContent);
 				require $fileName;
 			}
